requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
sqlparse
